pyproject.toml
[tool.poetry]
name = "srcgrph-blog-analysis"
version = "0.1.0"
description = ""
authors = ["Scott Bailey <csb5t@virginia.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
spacy = "^3.1.1"
ipykernel = "^6.2.0"
jupyter = "^1.0.0"
streamlit = "^0.87.0"
spacy-streamlit = "^1.0.2"
python-frontmatter = "^1.0.0"
textdescriptives = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
